RegisterUnivInfo 코틀린 변경 및 MVP 패턴 적용

+++ app/src/main/java/com/univreview/App.java
@@ -53,7 +53,7 @@ public class App extends MultiDexApplication {
     public static SharedPreferencesActivity pref;
     public static final Gson gson = new Gson();
     public static Picasso picasso;
-    public static int universityId;
+    public static Long universityId;
     public static Long userId;
     public static String userToken;
     public static String registrationId;
@@ -119,14 +119,14 @@ public class App extends MultiDexApplication {
 
     public static void setUniversityId(long universityId) {
         Logger.v("set universityId id: " + universityId);
-        App.universityId = (int) universityId;
+        App.universityId =  universityId;
         pref.savePreferences("universityId", App.universityId);
     }
 
     public static void init() {
         userToken = pref.getPreferences("userToken", null);
         userId = pref.getPreferences("userId", 0l);
-        universityId = pref.getPreferences("universityId", 0);
+        universityId = pref.getPreferences("universityId", 0l);
         Logger.v("user token: " + userToken);
         Logger.v("user id: " + userId);
         Logger.v("university id: " + universityId);
